added RestControllerAdvice, exception handler

+++ src/main/java/pl/bulandamichal/Pokedex/service/DeletePokemonService.java
@@ -2,6 +2,7 @@ package pl.bulandamichal.Pokedex.service;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import pl.bulandamichal.Pokedex.exceptions.MissingPokemonByIdException;
 import pl.bulandamichal.Pokedex.repository.PokemonRepository;
 
 @Service
@@ -11,6 +12,9 @@ public class DeletePokemonService {
     private final PokemonRepository pokemonRepository;
 
     public void deletePokemonById(Long id){
+        if (!pokemonRepository.existsById(id)){
+            throw new MissingPokemonByIdException(id);
+        }
         pokemonRepository.deleteById(id);
     }
 }
